refactor count method into a lib
