+++ services/PanelService.js
@@ -2,32 +2,21 @@
 
 var Panel = require('../models/panel');
 
-var count = function(query, callback) {
-    Panel.count(query, function(err, count) {
-        if (err) {
-            return callback(err, null);
-        }
-        var result = { result: false };
-        if (count) {
-            result.result = true;
-        }
-        return callback(null, result);
-    });
-};
+var count = require('../lib/queryUtil').count;
 
 module.exports = {
     isChildFI: function(activityYear, respondentId, callback) {
         var query = { 'activity_year': activityYear, 'respondent_id': respondentId, 'parent_name': { '$ne': '' }, 'other_lender_code': '0' };
-        count(query, callback);
+        count('panel', query, callback);
     },
 
     isRespondentMBS: function(activityYear, respondentId, callback) {
         var query = { 'activity_year': activityYear, 'respondent_id': respondentId, 'parent_name': { '$ne': '' }, 'other_lender_code': { '$in': ['1', '2', '3'] } };
-        count(query, callback);
+        count('panel', query, callback);
     },
 
     isValidControlNumber: function(activityYear, agencyCode, respondentId, callback) {
         var query = { 'activity_year': activityYear, 'agency_code': agencyCode, 'respondent_id': respondentId };
-        count(query, callback);
+        count('panel', query, callback);
     }
 };
