+++ test/services/PanelServiceSpec.js
@@ -1,12 +1,7 @@
 /*global describe:false, expect:false, it:false, beforeEach:false, afterEach:false*/
 'use strict';
 
-var mockgoose = require('mockgoose');
-var mongoose = require('mongoose');
-
 var PanelService = require('../../services/PanelService');
-var rewire = require('rewire');
-var count = rewire('../../services/PanelService').__get__('count');
 
 describe('PanelService', function() {
     describe('isChildFI', function() {
@@ -94,20 +89,6 @@ describe('PanelService', function() {
         it('should return a result of false if the agency code and respondent id are not a valid combination', function(done) {
             PanelService.isValidControlNumber('2013', '2', '0000000001', function(err, result) {
                 expect(result.result).to.be.false();
-                done();
-            });
-        });
-    });
-
-    describe('count', function() {
-        it('should return err when there is a mongo error', function(done) {
-            mockgoose.setMockReadyState(mongoose.connection, 0);
-
-            count({}, function(err, result) {
-                expect(err).to.have.property('name');
-                expect(err.name).to.be('MongoError');
-                expect(result).to.be.null();
-                mockgoose.setMockReadyState(mongoose.connection, 1);
                 done();
             });
         });
